requirements.txt
Flask==1.1.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==0.15.5
numpy==1.19.2
scipy==1.5.4
scikit-learn==0.24.0
matplotlib==3.1.3
pandas==1.1.2
seaborn==0.10.0
joblib==1.0.1
click==7.1.2
joblib==1.0.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
h5py==2.10.0
tensorflow==2.4.1
pillow==7.2.0
Keras==2.2.5